+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Piral Changelog
 
+## 1.6.2 (tbd)
+
+- Fixed `registerBreadcrumbs` in `piral-breadcrumbs` failing when no matcher was specified
+- Updated dependencies
+- Removed dependency to `path-to-regexp` package
+- Added `loading-pilets` and `loaded-pilets` events
+
 ## 1.6.1 (August 7, 2024)
 
 - Improved CLI flag validation and documentation (#705)